Refactor dev application
