+++ dev/manage.py
@@ -8,7 +8,7 @@ from django.core.management import execute_from_command_line
 
 
 def main() -> None:
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "s3ff_example.settings")
+    os.environ["DJANGO_SETTINGS_MODULE"] = "settings"
     execute_from_command_line(sys.argv)
