+++ dev/s3ff_dev/templates/base.html
@@ -1,0 +1,12 @@
+{% load static %}
+<!doctype html>
+<html lang="">
+<head>
+  <meta charset="utf-8">
+  <title>{% block title %}{% endblock %}</title>
+  {% block head %}{% endblock %}
+</head>
+<body>
+  {% block content %}{% endblock %}
+</body>
+</html>
