+++ dev/s3ff_dev/templates/s3ff_dev/resource_confirm_delete.html
@@ -7,11 +7,8 @@
 <h1>Delete Resource</h1>
 
 <form method="post">
-    {% csrf_token %}
-    <p>Are you sure you want to delete "{{ object }}"?</p>
-    <input type="submit" value="Confirm">
+  {% csrf_token %}
+  <p>Are you sure you want to delete "{{ object }}"?</p>
+  <input type="submit" value="Confirm">
 </form>
 {% endblock %}
-
-
-
