--- example/s3ff_example/core/templates/base.html
@@ -1,10 +1,0 @@
-{% load static %}
-<html>
-<head>
-    <title>{% block title %}{% endblock %}</title>
-    {% block head %}{% endblock %}
-</head>
-<body>
-    {% block content %}{% endblock %}
-</body>
-</html>
