--- example/s3ff_example/wsgi.py
@@ -1,9 +1,0 @@
-from __future__ import annotations
-
-import os
-
-from django.core.wsgi import get_wsgi_application
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "s3ff_example.settings")
-
-application = get_wsgi_application()
